extract event
