+++ src/app.rs
@@ -1,6 +1,6 @@
 use anyhow::{Context, Result};
 use futures::{channel::mpsc::Sender, SinkExt};
-use std::{fmt::Display, sync::Arc};
+use std::sync::Arc;
 
 use axum::{
     extract::{FromRef, State},
@@ -11,35 +11,9 @@ use axum::{
 };
 use axum_github_webhook_extract::{GithubEvent, GithubToken};
 use octocrab::params::repos::Reference;
-use serde::Deserialize;
 use tower_http::trace::TraceLayer;
 
-pub(crate) const ORG: &str = "scverse";
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(tag = "action", rename_all = "snake_case")]
-pub(crate) enum Event {
-    Enqueue(Enqueue),
-}
-
-#[derive(Debug, Clone, Deserialize)]
-pub(crate) struct Enqueue {
-    pub repo: String,
-    pub branch: String,
-    pub run_on: Option<String>,
-}
-
-impl From<Enqueue> for Event {
-    fn from(val: Enqueue) -> Self {
-        Event::Enqueue(val)
-    }
-}
-
-impl Display for Enqueue {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{ORG}/{}@{}", self.repo, self.branch)
-    }
-}
+use crate::event::{Enqueue, Event, ORG};
 
 #[derive(Debug, Clone)]
 struct AppState {
